Added back path for textual asm building. Fixed bug where shl and shr would use ecx instead of cl.
