+++ src/main.rs
@@ -1,5 +1,6 @@
 mod codegen;
 mod emit;
+mod emit_iced;
 mod lexer;
 mod parser;
 mod pretty;
@@ -48,6 +49,10 @@ struct Args {
     /// Use GCC for linking
     #[arg(long)]
     gcc: bool,
+
+    /// Use text based asm building instead of iced (Deprecated)
+    #[arg(long)]
+    no_iced: bool,
 
     #[arg(short = 'o', long)]
     output: Option<String>,
@@ -137,37 +142,19 @@ fn main() {
         .output
         .unwrap_or_else(|| path.with_extension("").to_string_lossy().to_string());
 
-    let obj = emit::emit_object(&code_ast).expect("iced obj");
-
-    if args.s {
-        let mut formatter = iced_x86::GasFormatter::new();
-        let ops = formatter.options_mut();
-        ops.set_number_base(iced_x86::NumberBase::Decimal);
-        let mut output = MyFormatterOutput::new();
-
-        println!("Generated ASM");
-        let asm = emit::get_instructions(&code_ast).unwrap();
-        for ins in asm {
-            output.vec.clear();
-            formatter.format(&ins, &mut output);
-            for (text, kind) in output.vec.iter() {
-                print!("{}", get_color(text.as_str(), *kind));
-            }
-            println!();
+    if args.no_iced {
+        let asm = emit::emit_program(&code_ast);
+        if args.s {
+            print!("Generated asm:\n\n{asm}");
         }
-    }
+        let asm_file = format!("{out_file}.s");
 
-    let obj_file = format!("{out_file}.o");
-    fs::write(&obj_file, &obj).expect("Failed to write object file");
+        fs::write(&asm_file, asm).expect("Failed to write assembly file");
 
-    if args.gcc {
         let status = std::process::Command::new("gcc")
-            .arg("-nostdlib")
-            .arg("-static")
-            .arg(&obj_file)
+            .arg(&asm_file)
             .arg("-o")
-            .arg(&out_file)
-            .arg("-Wl,-e,_start")
+            .arg(out_file.clone())
             .status()
             .expect("Failed to execute gcc");
 
@@ -175,14 +162,56 @@ fn main() {
             println!("Compilation failed with status: {status}");
             std::process::exit(1);
         }
+
+        fs::remove_file(&asm_file).expect("Failed to delete assembly file");
     } else {
-        let linker = Linker::new();
-        let args_vec = ["-o", &out_file, &obj_file, "--entry=_start"];
-        let parsed = WildArgs::parse(args_vec.iter()).expect("parse args");
-        let _ = libwild::setup_tracing(&parsed);
-        linker.run(&parsed).expect("link failed");
+        let obj = emit_iced::emit_object(&code_ast).expect("iced obj");
+
+        if args.s {
+            let mut formatter = iced_x86::GasFormatter::new();
+            let ops = formatter.options_mut();
+            ops.set_number_base(iced_x86::NumberBase::Decimal);
+            let mut output = MyFormatterOutput::new();
+
+            println!("Generated ASM");
+            let asm = emit_iced::get_instructions(&code_ast).unwrap();
+            for ins in asm {
+                output.vec.clear();
+                formatter.format(&ins, &mut output);
+                for (text, kind) in output.vec.iter() {
+                    print!("{}", get_color(text.as_str(), *kind));
+                }
+                println!();
+            }
+        }
+
+        let obj_file = format!("{out_file}.o");
+        fs::write(&obj_file, &obj).expect("Failed to write object file");
+
+        if args.gcc {
+            let status = std::process::Command::new("gcc")
+                .arg("-nostdlib")
+                .arg("-static")
+                .arg(&obj_file)
+                .arg("-o")
+                .arg(&out_file)
+                .arg("-Wl,-e,_start")
+                .status()
+                .expect("Failed to execute gcc");
+
+            if !status.success() {
+                println!("Compilation failed with status: {status}");
+                std::process::exit(1);
+            }
+        } else {
+            let linker = Linker::new();
+            let args_vec = ["-o", &out_file, &obj_file, "--entry=_start"];
+            let parsed = WildArgs::parse(args_vec.iter()).expect("parse args");
+            let _ = libwild::setup_tracing(&parsed);
+            linker.run(&parsed).expect("link failed");
+        }
+        fs::remove_file(&obj_file).ok();
     }
-    fs::remove_file(&obj_file).ok();
 
     if args.run {
         let run_status = std::process::Command::new(&out_file)
